+++ acorntypes/main.go
@@ -32,6 +32,14 @@ func (t Types) String() string {
 	return string(t)
 }
 
+// Type FontSize (string)
+type FontSize string
+
+// Return a string
+func (fs FontSize) String() string {
+	return string(fs)
+}
+
 // Type AcornComponents ([]string)
 type AcornComponents []string
 
@@ -97,6 +105,27 @@ type ImageParams struct {
 	WidthColumn *WidthColumn
 }
 
+// Type CouponParams
+type CouponParams struct {
+	Content string
+	Button  *ButtonParams
+	Styles  *CouponStyles
+}
+
+// Type PromoItems
+type PromoItems struct {
+	Promo       *PromoParams
+	Symbol      *PromoParams
+	Description *PromoParams
+}
+
+// Type PromoParams
+type PromoParams struct {
+	Value string
+	Color *Color
+	Size  *FontSize
+}
+
 // Type ColumnStyles
 type ColumnStyles struct {
 	Align     *Align
@@ -133,12 +162,18 @@ type LabelStyles struct {
 	Outlined  bool
 	Color     *Color
 	TextColor *Color
-	Type      *Types // Deprecated: Uses better Outlined
 }
 
 // Type TestimonialStyles
 type TestimonialStyles struct {
 	Image       string
 	Align       *Align
+	BorderColor *Color
+}
+
+// Type CouponStyles
+type CouponStyles struct {
+	Dashed      bool
+	Color       *Color
 	BorderColor *Color
 }